scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
feedparser==6.0.11

transformers==4.41.2
sentencepiece==0.2.0

schedule==1.2.2
python-dotenv==1.0.1
